Remove subinterfaces
